Fix tests and stop building container images
This will now only scan a container image if an image reference is
detected. This can happen if the `image_ref` parameter is passed or if
metadata from the [docker-metadata](https://github.com/equinixmetal-buildkite/docker-metadata-buildkite-plugin)
is detected.

Note that only the first image reference (tag) from the `docker-metadata`
plugin will be used, since it is assumed that all tags point to the same
image.

This also removes the ability for this plugin to build container images
if it detects a Dockerfile. This functionality should rather be left
outside of this plugin.

For better usability, this also uses the buildkite-agent command to push
a notification to users when both errors and successes happen (e.g. a
repo or image scan passed or failed).

Finally, this also extensively fixed the unit tests which were not stubbing the
docker commands correctly; Thus not actually doing relevant tests. This
now inserts mocks appropriately so we can test the behavior of the
plugin in a more extensive manner.

Signed-off-by: Juan Antonio Osorio <juan.osoriorobles@eu.equinix.com>