Move trivy plugin to the pre-artifact hook
This uses the pre-artifact hook since it'll for sure run after the
post-command one... The idea is that the docker-build plugin runs in
post-command (straight after the user has done something that might be
related for the build). Once that's done and before we do any artifact
upload we can do a container scan.

Signed-off-by: Juan Antonio Osorio <juan.osoriorobles@eu.equinix.com>