pre-checkout: Get trivy from GitHub.
This is my initial take on installing trivy from a GitHub release.
The script will first check if trivy can be found in any of the
directories listed in the PATH environment variable. If trivy
cannot be located in any of the PATH directories, the script will
do the following:

1. Determine the current OS and CPU
2. Download trivy's hashes and tar.gz from GitHub
3. Authenticate the trivy tar.gz using the hashes file
4. Move the trivy executable to either the current working
   directory, or to ${HOME}/bin
5. Write the executable's file path to stdout

The final move / rename logic is configurable. The user can
specify the "install-in-home-bin" parameter with a value of
"yes" or "true" (in upper or lowercase). In such a case,
the script moves the executable to ${HOME}/bin. If the
parameter is not specified, then the trivy executable is
moved to the current working directory.

The trivy executable's final file path is written to stdout.
This should probably change... but the buildkite documentation
is unclear about how one should pass data from one hook to
another. So... any ideas? :)