Update README and add Makefile for tests
The Makefile ensures we're able to run tests in a consistent manner in
the same way CI will run them. Thus helping us in developing in a more
predictable manner. This is the same Makefile as other plugins are using
(`docker-build` and `docker-metadata`).

With the new linting, we also needed to fix the README file to ensure
that we now adhere to the linting standards. e.g. all plugin options
need to be documented, and the examples need to match the current
version.

Signed-off-by: Juan Antonio Osorio <juan.osoriorobles@eu.equinix.com>