post-command: Execute trivy directly.
This commit updates "post-command" to run the trivy executable
directly instead of using a Docker container. The path to the
trivy executable is provided by the "pre-checkout" hook via
an environment variable.

Due to limitations with the "bats" testing framework, I set
the PATH environment variable to contain the trivy executable's
parent directory. This is, frankly, half-ass... but I am unsure
of a better alternative.

This commit also updates the post-command's tests to handle
these changes.