Output multiple annotations instead of appending
This enables the scan to output multiple annotations instead of
appending the output. This enables us to clearly portray to folks that
an fs scan failed while the image scan didn't or viceversa.

This also now uses a unique value for the annotation's context. One for the
filesystem scan (`trivy-fs-scan`) one for the container scan
(`trivy-container-scan`), and finally a catch-all context for "other"
failures or messages (`trivy-scan`).

For more info, look at the buildkite docs: https://buildkite.com/docs/agent/v3/cli-annotate

Signed-off-by: Juan Antonio Osorio <juan.osoriorobles@eu.equinix.com>