Add `release` Makefile target
This effectively does a release by doing e.g. `TAG=v1.15.2 make release`.

Under the hood, this will create a new branch with the same name as the
tag, edit the README.md with examples that follow the appropriate
version, tag the new commit, and finally, push both the branch and tag
to the repository.

Note that this also fixes the examples which were wrongly formatted.

Signed-off-by: Juan Antonio Osorio <juan.osoriorobles@eu.equinix.com>